requirements.txt
azure-storage
azure-servicebus
